Moved files stuff to fileWrapper
